start_transcription/requirements.txt
yt-dlp
google-cloud-storage
fake-useragent
functions-framework
google-api-python-client
google-cloud-storage
google-cloud-speech
google-cloud-translate
pymongo